feat: transactions filtered loaded
